memory saving patch from philmoz in http://chdk.setepontos.com/index.php?topic=650.msg70056#msg70056 - changed some struct values from int to short (conf.c, gui_menu.h, platform.h) - gui code cleanup plus some helper functions in gui_draw - changed palette display under Miscellaneous stuff menu to use the same UI as the color picker in Visual settings (better and consistent UI, plus removes a chunk of code) - packed version of font_8x16_uni.h default font, plus new tool to create packed version (this is done during the build so if changes are need to the font they can be done in font_8x16_uni.h). - Change 'strings' pointer table in lang.c to an offset table (halves table size). - Changes lang.c to use one buffer for the strings instead of allocating a new buffer for each string,

+++ trunk/core/gui_menu.h
@@ -38,7 +38,7 @@ typedef struct {
 
 typedef struct {
     char                symbol;
-    int                 title;
+    short               title;
     void                (*on_change)(unsigned int item);
     const CMenuItem     *menu;
 } CMenu;
